backend/power/requirements.txt
fastapi
uvicorn
pydantic
langchain
langchain-groq
langchain-core
langgraph
kafka-python
requests
httpx
aiofiles
python-multipart
python-dateutil
python-dotenv
pytest
pytest-asyncio
pytest-mock
structlog
prometheus-fastapi-instrumentator